requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.0
Authlib==0.14.1
autopep8==1.5.7
beautifulsoup4==4.8.2
bootstrap3-datetime==2.4
cairocffi==1.1.0
CairoSVG==1.0.22
certifi==2017.11.5
cffi==1.14.0
chardet==3.0.4
configparser==3.5.0
coverage==4.4.2
cryptography==2.9.2
cssselect2==0.2.1
cycler==0.10.0
distlib==0.3.1
Django==3.2
django-autocomplete-light==3.4.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-global-permissions==0.2.4
django-letsencrypt==2.0.1
django-logentry-admin==1.0.3
django-oauth-toolkit==1.0.0
django-polymorphic==2.1.2
django-tempus-dominus==5.1.2.13
django-tinymce==2.6.0
djangorestframework==3.12.4
djangorestframework-datatables==0.5.0
djangorestframework-simplejwt==4.7.1
ecdsa==0.15
enum34==1.1.6
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.5.0
gunicorn==19.6.0
html5lib==1.0.1
httplib2==0.17.3
idna==2.6
isort==4.3.21
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
matplotlib==3.3.4
mccabe==0.6.1
numexpr==2.7.3
numpy==1.20.0
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.44
openpyxl==3.0.7
packaging==16.8
pandas==1.2.2
pdfrw==0.4
Pillow==7.0.0
pipenv==2020.8.13
pkginfo==1.4.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.18
pycryptodome==3.4.7
pyflakes==1.6.0
PyJWT==2.1.0
pylint==2.5.0
pyparsing==2.1.10
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.9.4
pypi==2.1
python-dateutil==2.8.1
pytz==2017.3
qrcode==5.3
requests==2.23.0
rsa==4.0
scikit-learn==0.24.1
scipy==1.6.0
six==1.10.0
soupsieve==2.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tinycss2==0.6.1
toml==0.10.0
urllib3==1.22
virtualenv==20.0.31
virtualenv-clone==0.5.4
webencodings==0.5.1
wrapt==1.12.1
